Minor bug fix

+++ lthmi_nav/src/map_divider_htile.cpp
@@ -13,7 +13,7 @@ namespace lthmi_nav {
 
 class HorizTileMapDivider :  public MapDivider {
     public:
-        double half;
+        int half;
 
         HorizTileMapDivider() :
             MapDivider()
